move factories to individual components. move modules to their own file.
