+++ src/components/Box.js
@@ -1,29 +1,35 @@
 import React from 'react'
-import cx from '../transform'
+import { classnames } from '../transform'
 
 /*
   Creates an element that uses its unknown props as css class names.
   Defaults to a div.
  */
-export default function Box({
-  style,
-  className,
-  onClick,
-  title,
-  children,
-  is = 'div',
-  ...propClasses
-}) {
-  const classes = cx(propClasses, className)
-
-  const props = {
+export const factory = mapper =>
+  function Box({
     title,
     style,
     onClick,
+    className,
     children,
-    className: classes,
-    role: onClick ? 'button' : undefined
+    is = 'div',
+    ...propClasses
+  }) {
+    const classes = mapper(propClasses, className)
+    const Tag = is
+
+    const props = {
+      title,
+      style,
+      onClick,
+      children,
+      className: classes,
+      role: onClick ? 'button' : undefined
+    }
+
+    return <Tag {...props} />
   }
 
-  return React.createElement(is, props)
-}
+export const Box = factory(classnames)
+
+export default Box
