+++ src/components/Text.js
@@ -1,9 +1,16 @@
 import React from 'react'
-import cx from '../transform'
+import { classnames } from '../transform'
 
 /*
   Creates a span element that uses its unknown props as css class names.
  */
-export default function Text({ className, children, ...propClasses }) {
-  return <span className={cx(propClasses, className)} children={children} />
-}
+export const factory = mapper =>
+  function Text({ className, children, ...propClasses }) {
+    return (
+      <span className={mapper(propClasses, className)} children={children} />
+    )
+  }
+
+export const Text = factory(classnames)
+
+export default Text
